requirements.txt
scikit-learn
numpy
ruamel.yaml
pytorch
matplotlib
pandas
nltk